+++ apps/web/src/components/login-button/index.tsx
@@ -1,0 +1,1 @@
+export * from './LoginButton';
